+++ lib/player.js
@@ -5,13 +5,14 @@ var Player = {
 		socket.on('set_initials', function(data){
 			if (data && data != socket.initials){
 				Player.update_initials(socket, data);
-				Player.reset_score(socket);
+				Score.reset_score(socket);
 			}
 			socket.emit('player.renamed', socket.initials);
 		});
 
 		socket.emit('player.created');
 	},
+
 	// Broadcast
 	// broadcast who's online boards
 	emit_online_players: function() {
@@ -23,34 +24,6 @@ var Player = {
 			"Querying for online players",
 			function(results){
 				io.sockets.emit('online', results);
-			});
-	},
-
-	// broadcast high score boards for players
-	emit_high_scores: function() {
-		DB.db_call("SELECT `initials`, TIMESTAMPDIFF(SECOND, connected_at, last_ping) AS `score_seconds`" +
-			" FROM `players`" +
-			" ORDER BY `score_seconds` DESC LIMIT 10", [], "Querying for high scores", function(results){
-				io.sockets.emit('high_scores', results);
-			});
-	},
-
-	// Updates for player
-	// set 0 score (eg for new initials)
-	reset_score: function(socket) {
-		if (!socket || !socket.player_id){
-			Logger.message('Error resetting score: player_id is falsey.');
-			return;
-		}
-
-		DB.db_call("UPDATE `players` SET `connected_at` = CURRENT_TIMESTAMP, `last_ping` = CURRENT_TIMESTAMP WHERE `id` = ?",
-			[socket.player_id],
-			"Resetting score",
-			function(results){
-				if (results.changedRows > 0){
-					socket.start_time = Timer.now();
-					socket.emit('score_minutes', 0);
-				}
 			});
 	},
