+++ server.js
@@ -43,7 +43,7 @@ _prompt.get({properties: {
 
 	setInterval(function(){
 		Player.emit_online_players();
-		Player.emit_high_scores();
+		Score.emit_high_scores();
 		Team.emit_team_high_scores();
 	}, 5000);
